:writing_hand: update models encoders
